fix for bug 1328444?
git-svn-id: https://conky.svn.sourceforge.net/svnroot/conky/trunk/conky@353 7f574dfc-610e-0410-a909-a81674777703

+++ src/conky.h
@@ -159,8 +159,8 @@ struct information {
 	double uptime;
 
 	/* memory information in kilobytes */
-	unsigned int mem, memmax, swap, swapmax;
-	unsigned int bufmem, buffers, cached;
+	unsigned long mem, memmax, swap, swapmax;
+	unsigned long bufmem, buffers, cached;
 
 	unsigned int procs;
 	unsigned int run_procs;
